refactor: consolidate constants and validation into single files
This refactoring eliminates code duplication and improves maintainability
by establishing clear separation of concerns.

New Files:
- lib/constants.ts: Single source of truth for all shared constants
- Consolidated validation in lib/validation.ts

Changes:
- Removed duplicate IndexTagNames and StringObj definitions
- Updated all imports to use centralized locations
- Added backward-compatible re-exports

Benefits:
- Single source of truth prevents inconsistencies
- Clear file boundaries improve maintainability
- Zero breaking changes (backward compatible)

Testing:
✅ All 332 tests pass
✅ Type checking passes
✅ Build succeeds
