+++ lib/sitemap-item-stream.ts
@@ -1,12 +1,7 @@
 import { Transform, TransformOptions, TransformCallback } from 'node:stream';
 import { InvalidAttr } from './errors.js';
-import { SitemapItem, ErrorLevel, TagNames } from './types.js';
+import { SitemapItem, ErrorLevel, TagNames, StringObj } from './types.js';
 import { element, otag, ctag } from './sitemap-xml.js';
-
-export interface StringObj {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [index: string]: any;
-}
 
 /**
  * Builds an attributes object for XML elements from configuration object
