+++ lib/utils.ts
@@ -15,243 +15,15 @@ import { createInterface, Interface } from 'node:readline';
 import { URL } from 'node:url';
 import {
   SitemapItem,
-  ErrorLevel,
   SitemapItemLoose,
   EnumYesNo,
   Img,
   LinkItem,
   VideoItem,
-  isValidChangeFreq,
-  isValidYesNo,
-  isAllowDeny,
-  isPriceType,
-  isResolution,
-  NewsItem,
-  ErrorHandler,
 } from './types.js';
-import {
-  ChangeFreqInvalidError,
-  InvalidAttrValue,
-  InvalidNewsAccessValue,
-  InvalidNewsFormat,
-  InvalidVideoDescription,
-  InvalidVideoDuration,
-  InvalidVideoFormat,
-  InvalidVideoRating,
-  NoURLError,
-  NoConfigError,
-  PriorityInvalidError,
-  InvalidVideoTitle,
-  InvalidVideoViewCount,
-  InvalidVideoTagCount,
-  InvalidVideoCategory,
-  InvalidVideoFamilyFriendly,
-  InvalidVideoRestriction,
-  InvalidVideoRestrictionRelationship,
-  InvalidVideoPriceType,
-  InvalidVideoResolution,
-  InvalidVideoPriceCurrency,
-} from './errors.js';
-import { validators } from './types.js';
 
-function validate(
-  subject: NewsItem | VideoItem | NewsItem['publication'],
-  name: string,
-  url: string,
-  level: ErrorLevel
-): void {
-  Object.keys(subject).forEach((key): void => {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    const val = subject[key];
-    if (validators[key] && !validators[key].test(val)) {
-      if (level === ErrorLevel.THROW) {
-        throw new InvalidAttrValue(key, val, validators[key]);
-      } else {
-        console.warn(`${url}: ${name} key ${key} has invalid value: ${val}`);
-      }
-    }
-  });
-}
-
-function handleError(error: Error, level: ErrorLevel): void {
-  if (level === ErrorLevel.THROW) {
-    throw error;
-  } else if (level === ErrorLevel.WARN) {
-    console.warn(error.name, error.message);
-  }
-}
-/**
- * Verifies all data passed in will comply with sitemap spec.
- * @param conf Options to validate
- * @param level logging level
- * @param errorHandler error handling func
- */
-export function validateSMIOptions(
-  conf: SitemapItem,
-  level = ErrorLevel.WARN,
-  errorHandler: ErrorHandler = handleError
-): SitemapItem {
-  if (!conf) {
-    throw new NoConfigError();
-  }
-
-  if (level === ErrorLevel.SILENT) {
-    return conf;
-  }
-
-  const { url, changefreq, priority, news, video } = conf;
-
-  if (!url) {
-    errorHandler(new NoURLError(), level);
-  }
-
-  if (changefreq) {
-    if (!isValidChangeFreq(changefreq)) {
-      errorHandler(new ChangeFreqInvalidError(url, changefreq), level);
-    }
-  }
-
-  if (priority) {
-    if (!(priority >= 0.0 && priority <= 1.0)) {
-      errorHandler(new PriorityInvalidError(url, priority), level);
-    }
-  }
-
-  if (news) {
-    if (
-      news.access &&
-      news.access !== 'Registration' &&
-      news.access !== 'Subscription'
-    ) {
-      errorHandler(new InvalidNewsAccessValue(url, news.access), level);
-    }
-
-    if (
-      !news.publication ||
-      !news.publication.name ||
-      !news.publication.language ||
-      !news.publication_date ||
-      !news.title
-    ) {
-      errorHandler(new InvalidNewsFormat(url), level);
-    }
-
-    validate(news, 'news', url, level);
-    validate(news.publication, 'publication', url, level);
-  }
-
-  if (video) {
-    video.forEach((vid): void => {
-      if (vid.duration !== undefined) {
-        if (vid.duration < 0 || vid.duration > 28800) {
-          errorHandler(new InvalidVideoDuration(url, vid.duration), level);
-        }
-      }
-      if (vid.rating !== undefined && (vid.rating < 0 || vid.rating > 5)) {
-        errorHandler(new InvalidVideoRating(url, vid.title, vid.rating), level);
-      }
-
-      if (
-        typeof vid !== 'object' ||
-        !vid.thumbnail_loc ||
-        !vid.title ||
-        !vid.description
-      ) {
-        // has to be an object and include required categories https://support.google.com/webmasters/answer/80471?hl=en&ref_topic=4581190
-        errorHandler(new InvalidVideoFormat(url), level);
-      }
-
-      if (vid.title.length > 100) {
-        errorHandler(new InvalidVideoTitle(url, vid.title.length), level);
-      }
-
-      if (vid.description.length > 2048) {
-        errorHandler(
-          new InvalidVideoDescription(url, vid.description.length),
-          level
-        );
-      }
-
-      if (vid.view_count !== undefined && vid.view_count < 0) {
-        errorHandler(new InvalidVideoViewCount(url, vid.view_count), level);
-      }
-
-      if (vid.tag.length > 32) {
-        errorHandler(new InvalidVideoTagCount(url, vid.tag.length), level);
-      }
-
-      if (vid.category !== undefined && vid.category?.length > 256) {
-        errorHandler(new InvalidVideoCategory(url, vid.category.length), level);
-      }
-
-      if (
-        vid.family_friendly !== undefined &&
-        !isValidYesNo(vid.family_friendly)
-      ) {
-        errorHandler(
-          new InvalidVideoFamilyFriendly(url, vid.family_friendly),
-          level
-        );
-      }
-
-      if (vid.restriction) {
-        if (!validators.restriction.test(vid.restriction)) {
-          errorHandler(
-            new InvalidVideoRestriction(url, vid.restriction),
-            level
-          );
-        }
-        if (
-          !vid['restriction:relationship'] ||
-          !isAllowDeny(vid['restriction:relationship'])
-        ) {
-          errorHandler(
-            new InvalidVideoRestrictionRelationship(
-              url,
-              vid['restriction:relationship']
-            ),
-            level
-          );
-        }
-      }
-
-      // TODO price element should be unbounded
-      if (
-        (vid.price === '' && vid['price:type'] === undefined) ||
-        (vid['price:type'] !== undefined && !isPriceType(vid['price:type']))
-      ) {
-        errorHandler(
-          new InvalidVideoPriceType(url, vid['price:type'], vid.price),
-          level
-        );
-      }
-      if (
-        vid['price:resolution'] !== undefined &&
-        !isResolution(vid['price:resolution'])
-      ) {
-        errorHandler(
-          new InvalidVideoResolution(url, vid['price:resolution']),
-          level
-        );
-      }
-
-      if (
-        vid['price:currency'] !== undefined &&
-        !validators['price:currency'].test(vid['price:currency'])
-      ) {
-        errorHandler(
-          new InvalidVideoPriceCurrency(url, vid['price:currency']),
-          level
-        );
-      }
-
-      validate(vid, 'video', url, level);
-    });
-  }
-
-  return conf;
-}
+// Re-export validateSMIOptions from validation.ts for backward compatibility
+export { validateSMIOptions } from './validation.js';
 
 /**
  * Combines multiple streams into one
